Assets restructure
