+++ assets/test/expected/alpha_alpha_Fast_NoFilter.png
