+++ assets/test/expected/alpha_blue_Fast_NoFilter.png
