+++ assets/test/expected/alpha_texture_Fast_NoFilter.png
