+++ assets/test/expected/alpha_white_Fast_NoFilter.png
