+++ assets/test/expected/destination_over_alpha_Fast_NoFilter.png
