+++ assets/test/expected/destination_over_blue_Fast_NoFilter.png
