+++ assets/test/expected/destination_over_texture_Fast_NoFilter.png
