+++ assets/test/expected/destination_over_white_Fast_NoFilter.png
