+++ assets/test/expected/disjoint_debug_alpha_Fast_NoFilter.png
