+++ assets/test/expected/disjoint_debug_blue_Fast_NoFilter.png
