+++ assets/test/expected/disjoint_debug_texture_Fast_NoFilter.png
