+++ assets/test/expected/disjoint_debug_white_Fast_NoFilter.png
