+++ assets/test/expected/disjoint_over_alpha_Fast_NoFilter.png
