+++ assets/test/expected/disjoint_over_blue_Fast_NoFilter.png
