+++ assets/test/expected/disjoint_over_texture_Fast_NoFilter.png
