+++ assets/test/expected/disjoint_over_white_Fast_NoFilter.png
