+++ assets/test/expected/disjoint_under_alpha_Fast_NoFilter.png
