+++ assets/test/expected/disjoint_under_blue_Fast_NoFilter.png
