+++ assets/test/expected/disjoint_under_texture_Fast_NoFilter.png
