+++ assets/test/expected/disjoint_under_white_Fast_NoFilter.png
