+++ assets/test/expected/first_bottom_alpha_Fast_NoFilter.png
