+++ assets/test/expected/first_bottom_blue_Fast_NoFilter.png
