+++ assets/test/expected/first_bottom_texture_Fast_NoFilter.png
