+++ assets/test/expected/first_bottom_white_Fast_NoFilter.png
