+++ assets/test/expected/first_top_alpha_Fast_NoFilter.png
