+++ assets/test/expected/first_top_blue_Fast_NoFilter.png
