+++ assets/test/expected/first_top_texture_Fast_NoFilter.png
