+++ assets/test/expected/first_top_white_Fast_NoFilter.png
