+++ assets/test/expected/mask_top_alpha_Fast_NoFilter.png
