+++ assets/test/expected/mask_top_blue_Fast_NoFilter.png
