+++ assets/test/expected/mask_top_texture_Fast_NoFilter.png
