+++ assets/test/expected/mask_top_white_Fast_NoFilter.png
