+++ assets/test/expected/multiplicative_alpha_Fast_NoFilter.png
