+++ assets/test/expected/multiplicative_blue_Fast_NoFilter.png
