+++ assets/test/expected/multiplicative_texture_Fast_NoFilter.png
