+++ assets/test/expected/multiplicative_white_Fast_NoFilter.png
