+++ assets/test/expected/source_over_alpha_Fast_NoFilter.png
