+++ assets/test/expected/source_over_blue_Fast_NoFilter.png
