+++ assets/test/expected/source_over_texture_Fast_NoFilter.png
