+++ assets/test/expected/source_over_white_Fast_NoFilter.png
