+++ src/cli.rs
@@ -6,11 +6,9 @@ use pconvert_rust::blending::{
 };
 use pconvert_rust::constants;
 use pconvert_rust::errors::PConvertError;
-use pconvert_rust::parallelism::{ResultMessage, ThreadPool};
-use pconvert_rust::utils::{read_png_from_file, write_png_parallel, write_png_to_file};
+use pconvert_rust::tests::{compose, compose_parallel, Background};
+use pconvert_rust::utils::read_png_from_file;
 use std::env;
-use std::fmt;
-use std::fmt::{Display, Formatter};
 use std::str;
 use std::str::FromStr;
 
@@ -179,183 +177,4 @@ fn apply_blue_filter(pixel: &mut Rgba<u8>) {
     // sets red value to 0 and green value to the blue one (blue filter effect)
     pixel[0] = 0;
     pixel[1] = pixel[2];
-}
-
-fn compose(
-    dir: &str,
-    algorithm: BlendAlgorithm,
-    background: Background,
-    compression: CompressionType,
-    filter: FilterType,
-    benchmark: &mut Benchmark,
-) -> Result<(), PConvertError> {
-    let demultiply = is_algorithm_multiplied(&algorithm);
-
-    let mut bot = benchmark.execute(Benchmark::add_read_png_time, || {
-        read_png_from_file(format!("{}sole.png", dir), demultiply)
-    })?;
-
-    let algorithm_fn = get_blending_algorithm(&algorithm);
-
-    // reads one PNG at the time and blends it with the current result
-
-    let top = benchmark.execute(Benchmark::add_read_png_time, || {
-        read_png_from_file(format!("{}back.png", dir), demultiply)
-    })?;
-
-    benchmark.execute(Benchmark::add_blend_time, || {
-        blend_images(&top, &mut bot, &algorithm_fn, &None)
-    });
-
-    let top = benchmark.execute(Benchmark::add_read_png_time, || {
-        read_png_from_file(format!("{}front.png", dir), demultiply)
-    })?;
-
-    benchmark.execute(Benchmark::add_blend_time, || {
-        blend_images(&top, &mut bot, &algorithm_fn, &None)
-    });
-
-    let top = benchmark.execute(Benchmark::add_read_png_time, || {
-        read_png_from_file(format!("{}shoelace.png", dir), demultiply)
-    })?;
-
-    benchmark.execute(Benchmark::add_blend_time, || {
-        blend_images(&top, &mut bot, &algorithm_fn, &None)
-    });
-
-    if demultiply {
-        benchmark.execute(Benchmark::add_blend_time, || multiply_image(&mut bot));
-    }
-
-    let mut composition = benchmark.execute(Benchmark::add_read_png_time, || {
-        read_png_from_file(format!("{}background_{}.png", dir, background), false)
-    })?;
-
-    benchmark.execute(Benchmark::add_blend_time, || {
-        blend_images(&bot, &mut composition, &algorithm_fn, &None)
-    });
-
-    // writes the final composition to the file system
-    let file_out = format!(
-        "{}result_{}_{}_{:#?}_{:#?}.png",
-        dir, algorithm, background, compression, filter
-    );
-    benchmark.execute(Benchmark::add_write_png_time, || {
-        write_png_to_file(file_out, &composition, compression, filter)
-    })?;
-
-    Ok(())
-}
-
-fn compose_parallel(
-    dir: &str,
-    algorithm: BlendAlgorithm,
-    background: Background,
-    compression: CompressionType,
-    filter: FilterType,
-    benchmark: &mut Benchmark,
-) -> Result<(), PConvertError> {
-    let demultiply = is_algorithm_multiplied(&algorithm);
-    let algorithm_fn = get_blending_algorithm(&algorithm);
-
-    let mut thread_pool = ThreadPool::new(5)?;
-
-    // sends the PNG reading tasks to multiple threads
-    // these values are hardcoded by the multiple layer files
-    let png_file_names = vec![
-        "sole.png".to_owned(),
-        "back.png".to_owned(),
-        "front.png".to_owned(),
-        "shoelace.png".to_owned(),
-        format!("background_{}.png", background),
-    ];
-
-    let mut result_channels = Vec::with_capacity(png_file_names.len());
-    thread_pool.start();
-    for png_file_name in png_file_names {
-        let path = format!("{}{}", dir, png_file_name);
-        let result_channel = thread_pool
-            .execute(move || ResultMessage::ImageResult(read_png_from_file(path, demultiply)));
-        result_channels.push(result_channel);
-    }
-
-    // blending phase, will run the multiple layers operation
-    // as expected by the proper execution
-    let mut bot = benchmark.execute(Benchmark::add_read_png_time, || {
-        match result_channels[0].recv().unwrap() {
-            ResultMessage::ImageResult(result) => result,
-        }
-    })?;
-
-    let top = benchmark.execute(Benchmark::add_read_png_time, || {
-        match result_channels[1].recv().unwrap() {
-            ResultMessage::ImageResult(result) => result,
-        }
-    })?;
-    benchmark.execute(Benchmark::add_blend_time, || {
-        blend_images(&top, &mut bot, &algorithm_fn, &None)
-    });
-
-    let top = benchmark.execute(Benchmark::add_read_png_time, || {
-        match result_channels[2].recv().unwrap() {
-            ResultMessage::ImageResult(result) => result,
-        }
-    })?;
-    benchmark.execute(Benchmark::add_blend_time, || {
-        blend_images(&top, &mut bot, &algorithm_fn, &None)
-    });
-
-    let top = benchmark.execute(Benchmark::add_read_png_time, || {
-        match result_channels[3].recv().unwrap() {
-            ResultMessage::ImageResult(result) => result,
-        }
-    })?;
-    benchmark.execute(Benchmark::add_blend_time, || {
-        blend_images(&top, &mut bot, &algorithm_fn, &None)
-    });
-
-    if demultiply {
-        multiply_image(&mut bot);
-    }
-
-    let mut composition =
-        benchmark.execute(Benchmark::add_read_png_time, || {
-            match result_channels[4].recv().unwrap() {
-                ResultMessage::ImageResult(result) => result,
-            }
-        })?;
-    benchmark.execute(Benchmark::add_blend_time, || {
-        blend_images(&bot, &mut composition, &algorithm_fn, &None)
-    });
-
-    // writes the final composition PNG to the output file,
-    // this is considered to be the most expensive operation
-    let file_out = format!(
-        "{}result_{}_{}_{:#?}_{:#?}.png",
-        dir, algorithm, background, compression, filter
-    );
-    benchmark.execute(Benchmark::add_write_png_time, || {
-        write_png_parallel(file_out, &composition, compression, filter)
-    })?;
-
-    Ok(())
-}
-
-#[derive(Clone)]
-pub enum Background {
-    Alpha,
-    White,
-    Blue,
-    Texture,
-}
-
-impl Display for Background {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        match self {
-            Background::Alpha => write!(f, "alpha"),
-            Background::White => write!(f, "white"),
-            Background::Blue => write!(f, "blue"),
-            Background::Texture => write!(f, "texture"),
-        }
-    }
 }
